requirements.txt
absl-py==1.4.0
altair==5.0.1
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
executing==1.2.0
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.3
ipython==8.14.0
itsdangerous==2.1.2
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.8.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.24.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
yarg==0.1.9
zipp==3.15.0
